Adressing more feedback
